Update ViewController.swift
Line 2744 is forcefully unwrapping the value from the dictionary (xmlPrefOptions). We are seeing crashes where the key (saveRawXmlScope) does not exist. Ideally should check if a value is returned before unwrapping. Here I have used the Nil Coalescing Operator to supply a default value of false if the key is not found. You could have used an if/let statement as well.